restructure shims
