+++ src/cli/javascript/mod.rs
@@ -1,6 +1,6 @@
 mod http_filter;
 mod runtime;
 mod serde_v8;
-
+mod shim;
 pub use http_filter::HttpFilter;
 pub use runtime::Runtime;
